Moved compression to separate feature
