+++ src/modules/Elsa.Common/Codecs/GZip.cs
@@ -1,7 +1,7 @@
 using System.IO.Compression;
 using System.Text;
 
-namespace Elsa.Workflows.Management.Compression;
+namespace Elsa.Common.Codecs;
 
 /// <summary>
 /// Represents a GZip compression strategy.
