+++ src/modules/Elsa.Common/Codecs/None.cs
@@ -1,4 +1,4 @@
-namespace Elsa.Workflows.Management.Compression;
+namespace Elsa.Common.Codecs;
 
 /// <summary>
 /// Represents a compression strategy that does not compress or decompress the input.
